+++ src/index.ts
@@ -1,8 +1,16 @@
 export { MD_Collection, type MD_Collection_Parameter_Type } from "./md-collection";
-export { MD_Frontmatter, type MD_Frontmatter_Type } from "./md-frontmatter";
+export { MD_CopyJob } from "./md-copy-job";
+export { MD_Exporter, type MD_Exporter_Parameter_Type } from  "./md-exporter";
+export { MD_Filesystem } from "./md-filesystem";
+export { MD_Frontmatter_Template, type MD_Frontmatter_Type } from "./md-frontmatter";
+export { MD_ObserverSubject, type MD_Observer_Interface } from "./md-observer";
 export { MD_Template } from "./md-template";
 
+export { MD_Transformer_Factory } from "./md-transformer-factory";
+
 export {
+  MD_Transformer_AbstractBase,
+  type MD_CopyTask_Type,
   type MD_Transformer_Interface,
   type MD_Transformer_Parameter_Type,
   type MD_Transformer_TemplateValues_Type