+++ test/md-export-3.test.ts
@@ -1,25 +1,27 @@
-import { MD_CopyTask_Type } from './../src/md-transformer';
-import { MD_Frontmatter } from "../src/md-frontmatter";
+import { MD_CopyTask_Type } from '../src/md-transformer';
+import { MD_Frontmatter_Template } from "../src/md-frontmatter";
 import { MD_Exporter, MD_Exporter_Parameter_Type } from "../src/md-exporter";
 import { MD_Transformer_Parameter_Type } from "../src/md-transformer";
-import { MD_Splitter_Transformer } from "../src/transformer/md-splitter-task";
 import { MD_Splitter_Parameter_Type } from "../src/transformer/md-splitter-task";
 import { MD_ObsidianLink_Transformer } from "../src/transformer/md-obsidian-link-task";
 import { MD_RemoveTODOS_Transformer } from "../src/transformer/md-remove-todos-task";
+import { MD_Frontmatter_Transformer } from '../src/transformer/md-frontmatter-task';
 
 const exporter: MD_Exporter = new MD_Exporter();
 
 // Basic instructions for MD_Exporter
 
+const simulate_job = false;
+const simulate_copy_job = false;
+
 const exporter_parameter: MD_Exporter_Parameter_Type = {
-  readPath: "test/obsidian-vault/longform.md",
-  writePath: "test/hugo-content-2/",
+  readPath: "test/obsidian-vault/some-md-docs",
+  writePath: "test/hugo-content-3/",
   doSubfolders: false,
   limit: 1990,
-  useCounter: false
+  useCounter: false,
+  simulate: simulate_job
 };
-
-const simulate_copy_job = true;
 
 // The tasks to operate. If you add not tasks you have a simple copy job.
 
@@ -30,7 +32,7 @@ const parameter_images: MD_Transformer_Parameter_Type = {
   replace_template: `{{< image src="assets/images/{name_full}" >}}`,
   copy_task: {
     source:"test/obsidian-vault/images/",
-    target:"test/hugo-content-2/assets/images/{name}/",
+    target:"test/hugo-content-3/assets/images/{name}/",
     simulate:simulate_copy_job
   }
 };
@@ -42,7 +44,7 @@ const parameter_docs: MD_Transformer_Parameter_Type = {
   replace_template: `{{< button href="/getthis.php?id={name}" name="download {name} ({name_suffix})" >}}`,
   copy_task: {
     source:"test/obsidian-vault/attachments/",
-    target:"test/hugo-content-2/static/downloads/",
+    target:"test/hugo-content-3/static/downloads/",
     simulate:simulate_copy_job
   }
 };
@@ -58,7 +60,7 @@ const parameter_remove: MD_Transformer_Parameter_Type = {
 
 // The placeholders result from the definition of MD_Frontmatter_Type in md-frontmatter
 // A file definition in frontmatter_filename, overwrites frontmatter Property
-var splitter_frontmatter: MD_Frontmatter = new MD_Frontmatter(`---
+var document_frontmatter: MD_Frontmatter_Template = new MD_Frontmatter_Template(`---
 title: "{title}"
 description: ""
 url: /docs/{url_prefix}-{url}/
@@ -82,12 +84,12 @@ const parameter_splitter: MD_Splitter_Parameter_Type = {
   url_prefix: "test-prefix",
   doRemoveHeadline: true,
   frontmatter_filename: "", // ./test/frontmatter-template.md
-  frontmatter: splitter_frontmatter
+  frontmatter: document_frontmatter
 };
 
 exporter.addTransformer(new MD_ObsidianLink_Transformer(parameter_images));
 exporter.addTransformer(new MD_ObsidianLink_Transformer(parameter_docs));
 exporter.addTransformer(new MD_RemoveTODOS_Transformer(parameter_remove));
-exporter.addTransformer(new MD_Splitter_Transformer(parameter_splitter));
+exporter.addTransformer(new MD_Frontmatter_Transformer(parameter_splitter));
 
 exporter.perform_job(exporter_parameter);